catalogs/uniprot/requirements.txt
biopython==1.79
regex==2021.8.28
